+++ NEWS.md
@@ -8,16 +8,44 @@
 * Input checks on multiple functions
 
 # rotsee UNRELEASED
+# rotsee v0.1.0 2025-09-24
+## added
+* B_LU_START and B_LU_END as input parameters in rothc_rotation
+* start_date and end_date to replace simyears
+* Helper functions rc_extend_crops and rc_extend_amendments to extend user input tables
+* Helper function rc_time_period to generate a table of all years and months in the simulation period
+
+## changed
+* Base soil cover on user supplied crop growing dates
+* Calculate actual evapotranspiration based on simple rothc correction factor
+* Corrected accumulated soil moisture deficit calculation
+
+# rotsee v0.0.4 2025-09-24
+## Changed
+* Split `rc_input_events` into `rc_input_event_crop`, `rc_input_event_amendment`, and merge these in `rc_input_events`
+* Added unit tests for `rc_input_event_crop`, `rc_input_event_amendment`, and `rc_input_events`
+* Expand README with some introductory text about the package
+  
+# rotsee v0.0.3 2025-09-01
+## Added
+* function rc_check_inputs to check input data
+* Helper function rc_calculate_bd to estimate dry soil bulk density based on soil properties
+* Helper function rc_calculate_B_C_OF to calculate crop C inputs based on crop management
+
+## changed
+* Input data of crops and amendments to be optional
+* Option to input total crop C input or general crop management data
+* Changed amendment inputs to single event date, format yyyy-mm-dd
+
+# rotsee V0.0.2 2025-08-28
+
 ## Added
 * Add GitHub Action to run R-CMD-CHECK for PR's
 * Added helper functions to check input weather and parameter data, and insert default values when not supplied
 * Added unit tests for helper functions
 
-## Changed
-* Expand README with some introductory text about the package
-
 ## Fixed
-* Add missing dependenies for `roxygen2`, `devtools` and `usethis`
+* Add missing dependencies for `roxygen2`, `devtools` and `usethis`
 
 # rotsee v0.0.1 2025-07-31